GET /api/restaurants
