+++ controllers/api/restController.js
@@ -1,0 +1,11 @@
+const restService = require('../../services/restService');
+
+const restController = {
+  getRestaurants: (req, res) => {
+    restService.getRestaurants(req, res, data => {
+      return res.json(data);
+    });
+  }
+};
+
+module.exports = restController;
