+++ controllers/restController.js
@@ -1,3 +1,5 @@
+const restService = require('../services/restService');
+
 const db = require('../models');
 const Sequelize = require('sequelize');
 const Restaurant = db.Restaurant;
@@ -10,51 +12,8 @@ const pageLimit = 10;
 
 const restController = {
   getRestaurants: (req, res) => {
-    let offset = 0;
-    let whereQuery = {};
-    let categoryId = '';
-
-    if (req.query.page) {
-      offset = (req.query.page - 1) * pageLimit;
-    }
-
-    if (req.query.categoryId) {
-      categoryId = Number(req.query.categoryId);
-      whereQuery['CategoryId'] = categoryId;
-    }
-
-    Restaurant.findAndCountAll({
-      include: Category,
-      where: whereQuery,
-      offset: offset,
-      limit: pageLimit
-    }).then(result => {
-      let page = Number(req.query.page) || 1;
-      let pages = Math.ceil(result.count / pageLimit);
-      let totalPage = Array.from({ length: pages }).map(
-        (item, index) => index + 1
-      );
-      let prev = page - 1 < 1 ? 1 : page - 1;
-      let next = page + 1 > pages ? pages : page + 1;
-      const data = result.rows.map(r => ({
-        ...r.dataValues,
-        description: r.dataValues.description.substring(0, 50),
-        isFavorited: req.user.FavoritedRestaurants.map(d => d.id).includes(
-          r.id
-        ),
-        isLiked: req.user.LikedRestaurants.map(d => d.id).includes(r.id)
-      }));
-      Category.findAll().then(categories => {
-        return res.render('restaurants', {
-          restaurants: data,
-          categories,
-          categoryId,
-          page,
-          totalPage,
-          prev,
-          next
-        });
-      });
+    restService.getRestaurants(req, res, data => {
+      return res.render('restaurants', data);
     });
   },
